+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #47 Set<Package> bound in JerseyModule must be qualified with annotation
 * #48 Combine bootique-jersey with bootique-jersey-client in one project 
 * #49 Upgrade to Jersey 2.30.1
+* #50 Integrate 'jersey-bean-validation'
 
 ## 1.0.RC1